Move interface to spi package
