+++ src/main/java/no/entur/uttu/graphql/mappers/LineMapper.java
@@ -5,7 +5,7 @@ import static no.entur.uttu.graphql.GraphQLNames.FIELD_NOTICES;
 
 import no.entur.uttu.graphql.ArgumentWrapper;
 import no.entur.uttu.model.Line;
-import no.entur.uttu.organisation.OrganisationRegistry;
+import no.entur.uttu.organisation.spi.OrganisationRegistry;
 import no.entur.uttu.repository.NetworkRepository;
 import no.entur.uttu.repository.ProviderRepository;
 import no.entur.uttu.repository.generic.ProviderEntityRepository;
