+++ src/main/java/no/entur/uttu/graphql/mappers/ServiceJourneyMapper.java
@@ -20,7 +20,7 @@ import static no.entur.uttu.graphql.GraphQLNames.*;
 import no.entur.uttu.graphql.ArgumentWrapper;
 import no.entur.uttu.model.DayType;
 import no.entur.uttu.model.ServiceJourney;
-import no.entur.uttu.organisation.OrganisationRegistry;
+import no.entur.uttu.organisation.spi.OrganisationRegistry;
 import no.entur.uttu.repository.ProviderRepository;
 import no.entur.uttu.repository.generic.ProviderEntityRepository;
 import org.springframework.stereotype.Component;
